refactor: improve pkce handling
