+++ src/server/index.js
@@ -12,7 +12,7 @@ import * as routes from './routes'
 import renderPage from './pages'
 import csrfTokenHandler from './lib/csrf-token-handler'
 import createSecret from './lib/create-secret'
-import * as pkce from './lib/pkce-handler'
+import * as pkce from './lib/oauth/pkce-handler'
 
 // To work properly in production with OAuth providers the NEXTAUTH_URL
 // environment variable must be set.
@@ -145,9 +145,7 @@ async function NextAuthHandler (req, res, userOptions) {
           return render.signout()
         case 'callback':
           if (provider) {
-            const error = await pkce.handleCallback(req, res)
-            if (error) return res.redirect(error)
-
+            if (await pkce.handleCallback(req, res)) return
             return routes.callback(req, res)
           }
           break
@@ -184,9 +182,7 @@ async function NextAuthHandler (req, res, userOptions) {
         case 'signin':
           // Verified CSRF Token required for all sign in routes
           if (csrfTokenVerified && provider) {
-            const error = await pkce.handleSignin(req, res)
-            if (error) return res.redirect(error)
-
+            if (await pkce.handleSignin(req, res)) return
             return routes.signin(req, res)
           }
 
@@ -204,9 +200,7 @@ async function NextAuthHandler (req, res, userOptions) {
               return res.redirect(`${baseUrl}${basePath}/signin?csrf=true`)
             }
 
-            const error = await pkce.handleCallback(req, res)
-            if (error) return res.redirect(error)
-
+            if (await pkce.handleCallback(req, res)) return
             return routes.callback(req, res)
           }
           break
