+++ src/server/lib/oauth/pkce-handler.js
@@ -1,7 +1,8 @@
 import pkceChallenge from 'pkce-challenge'
-import jwt from '../../lib/jwt'
-import * as cookie from '../lib/cookie'
-import logger from '../../lib/logger'
+import * as cookie from '../cookie'
+import jwt from '../../../lib/jwt'
+import logger from '../../../lib/logger'
+import { OAuthCallbackError } from '../../../lib/errors'
 
 const PKCE_LENGTH = 64
 const PKCE_CODE_CHALLENGE_METHOD = 'S256' // can be 'plain', not recommended https://tools.ietf.org/html/rfc7636#section-4.2
@@ -16,7 +17,7 @@ export async function handleCallback (req, res) {
     }
 
     if (!(cookies.pkceCodeVerifier.name in req.cookies)) {
-      throw new Error('The code_verifier cookie was not found.')
+      throw new OAuthCallbackError('The code_verifier cookie was not found.')
     }
     const pkce = await jwt.decode({
       ...req.options.jwt,
@@ -24,11 +25,16 @@ export async function handleCallback (req, res) {
       maxAge: PKCE_MAX_AGE,
       encryption: true
     })
-    cookie.set(res, cookies.pkceCodeVerifier.name, null, { maxAge: 0 }) // remove PKCE after it has been used
     req.options.pkce = pkce
+    logger.debug('OAUTH_CALLBACK_PROTECTION', 'Read PKCE verifier from cookie', {
+      code_verifier: pkce.code_verifier,
+      pkceLength: PKCE_LENGTH,
+      method: PKCE_CODE_CHALLENGE_METHOD
+    })
+    cookie.set(res, cookies.pkceCodeVerifier.name, null, { maxAge: 0 }) // remove PKCE after it has been used
   } catch (error) {
-    logger.error('PKCE_ERROR', error)
-    return `${baseUrl}${basePath}/error?error=Configuration`
+    logger.error('CALLBACK_OAUTH_ERROR', error)
+    return res.redirect(`${baseUrl}${basePath}/error?error=OAuthCallback`)
   }
 }
 
@@ -41,10 +47,18 @@ export async function handleSignin (req, res) {
     }
     // Started login flow, add generated pkce to req.options and (encrypted) code_verifier to a cookie
     const pkce = pkceChallenge(PKCE_LENGTH)
-    req.options.pkce = {
+    logger.debug('OAUTH_SIGNIN_PROTECTION', 'Created PKCE challenge/verifier', {
+      ...pkce,
+      pkceLength: PKCE_LENGTH,
+      method: PKCE_CODE_CHALLENGE_METHOD
+    })
+
+    provider.authorizationParams = {
+      ...provider.authorizationParams,
       code_challenge: pkce.code_challenge,
       code_challenge_method: PKCE_CODE_CHALLENGE_METHOD
     }
+
     const encryptedCodeVerifier = await jwt.encode({
       ...req.options.jwt,
       maxAge: PKCE_MAX_AGE,
@@ -58,12 +72,11 @@ export async function handleSignin (req, res) {
       expires: cookieExpires.toISOString(),
       ...cookies.pkceCodeVerifier.options
     })
+    logger.debug('OAUTH_SIGNIN_PROTECTION', 'Created PKCE code_verifier saved in cookie')
   } catch (error) {
-    logger.error('PKCE_ERROR', error)
-    return `${baseUrl}${basePath}/error?error=Configuration`
+    logger.error('SIGNIN_OAUTH_ERROR', error)
+    return res.redirect(`${baseUrl}${basePath}/error?error=OAuthSignin`)
   }
 }
 
-export default {
-  handleSignin, handleCallback
-}
+export default { handleSignin, handleCallback }
